Lego Color Visual Table
The table created represents Lego brick's solid colors that are currently available in the market. This table was designed for potential buyers and Lego collectors to help visualize them information regarding the Lego brick's solid colors that are currently available and the rarity for each solid color. The list of solid colors excludes previous solid colors that their years has ended as they are believed to not be currently available. Solid colors that their year-end are uncertain/unknown are also being excluded from the table. This leaves in the table a list of solid colors that are confirmed to be currently available. Some of the colors are seen as limited edition. This type of information may be useful to collectors so they may be aware of what solid color bricks are currently limited in Lego’s current market. As for buyers in general, this table may be useful for them to simply look up at the list knowing what solid colors overall are currently available for them to purchase.